tile component refactor

+++ src/lib/jp2/codestream/Quantizer.cpp
@@ -28,7 +28,7 @@ void Quantizer::setBandStepSizeAndBps(TileCodingParams *tcp,
 		                   uint32_t resno,
 						   uint8_t bandno,
 							TileComponentCodingParams *tccp,
-							uint32_t image_precision,
+							uint8_t image_precision,
 							bool compress){
 
 	uint32_t gain = 0;
